requirements.txt
colorama==0.4.5
pycryptodome==3.17
